retire locations.getExplorerData from interface
