+++ core/src/api/mod.rs
@@ -1,21 +1,9 @@
-use chrono::{DateTime, Utc};
-use prisma_client_rust::{operator::or, Direction};
 use rspc::{alpha::Rspc, Config};
 use serde::{Deserialize, Serialize};
 use specta::Type;
 use std::sync::Arc;
-use uuid::Uuid;
 
-use crate::{
-	api::{
-		locations::{file_path_with_object, ExplorerItem},
-		utils::library,
-	},
-	location::LocationError,
-	node::NodeConfig,
-	prisma::*,
-	Node,
-};
+use crate::{node::NodeConfig, Node};
 
 use utils::{InvalidRequests, InvalidateOperationEvent};
 
@@ -39,6 +27,7 @@ mod libraries;
 mod locations;
 mod nodes;
 mod p2p;
+mod search;
 mod sync;
 mod tags;
 pub mod utils;
@@ -80,127 +69,7 @@ pub(crate) fn mount() -> Arc<Router> {
 				})
 			})
 		})
-		.procedure("search", {
-			#[derive(Deserialize, Type, Debug, Clone, Copy)]
-			#[serde(rename_all = "camelCase")]
-			#[specta(inline)]
-			enum Ordering {
-				Name(bool),
-			}
-			impl Ordering {
-				fn get_direction(&self) -> Direction {
-					match self {
-						Self::Name(v) => v,
-					}
-					.then_some(Direction::Asc)
-					.unwrap_or(Direction::Desc)
-				}
-				fn to_param(self) -> file_path::OrderByParam {
-					let dir = self.get_direction();
-					use file_path::*;
-					match self {
-						Self::Name(_) => name::order(dir),
-					}
-				}
-			}
-			#[derive(Deserialize, Type, Debug)]
-			#[serde(rename_all = "camelCase")]
-			#[specta(inline)]
-			struct Args {
-				#[specta(optional)]
-				location_id: Option<i32>,
-				#[specta(optional)]
-				after_file_id: Option<Uuid>,
-				#[specta(optional)]
-				take: Option<i32>,
-				#[specta(optional)]
-				order: Option<Ordering>,
-				#[specta(optional)]
-				search: Option<String>,
-				#[specta(optional)]
-				extension: Option<String>,
-				#[specta(optional)]
-				kind: Option<i32>,
-				#[specta(optional)]
-				#[serde(default)]
-				tags: Vec<i32>,
-				#[specta(optional)]
-				created_at_from: Option<DateTime<Utc>>,
-				#[specta(optional)]
-				created_at_to: Option<DateTime<Utc>>,
-				#[specta(optional)]
-				path: Option<String>,
-			}
-
-			R.with2(library())
-				.query(|(_, library), args: Args| async move {
-					let params = args
-						.search
-						.map(|search| {
-							search
-								.split(' ')
-								.map(str::to_string)
-								.map(file_path::materialized_path::contains)
-								.map(Some)
-								.collect::<Vec<_>>()
-						})
-						.unwrap_or_default()
-						.into_iter()
-						.chain([
-							args.location_id.map(file_path::location_id::equals),
-							args.kind.map(|kind| {
-								file_path::object::is(vec![object::kind::equals(kind)])
-							}),
-							args.extension.map(file_path::extension::equals),
-							(!args.tags.is_empty()).then(|| {
-								file_path::object::is(vec![object::tags::some(vec![
-									tag_on_object::tag::is(vec![or(args
-										.tags
-										.into_iter()
-										.map(tag::id::equals)
-										.collect())]),
-								])])
-							}),
-							args.created_at_from
-								.map(|v| file_path::date_created::gte(v.into())),
-							args.created_at_to
-								.map(|v| file_path::date_created::lte(v.into())),
-							args.path.map(file_path::materialized_path::starts_with),
-						])
-						.flatten()
-						.collect();
-					let mut query = library.db.file_path().find_many(params);
-					if let Some(file_id) = args.after_file_id {
-						query = query.cursor(file_path::pub_id::equals(file_id.as_bytes().to_vec()))
-					}
-					if let Some(order) = args.order {
-						query = query.order_by(order.to_param());
-					}
-					if let Some(take) = args.take {
-						query = query.take(take as i64);
-					}
-					let file_paths = query
-						.include(file_path_with_object::include())
-						.exec()
-						.await?;
-					let mut items = Vec::with_capacity(file_paths.len());
-					for file_path in file_paths {
-						let has_thumbnail = if let Some(cas_id) = &file_path.cas_id {
-							library
-								.thumbnail_exists(cas_id)
-								.await
-								.map_err(LocationError::IOError)?
-						} else {
-							false
-						};
-						items.push(ExplorerItem::Path {
-							has_thumbnail,
-							item: file_path,
-						})
-					}
-					Ok(items)
-				})
-		})
+		.merge("search.", search::mount())
 		.merge("library.", libraries::mount())
 		.merge("volumes.", volumes::mount())
 		.merge("tags.", tags::mount())
