+++ interface/app/$libraryId/Explorer/util.ts
@@ -10,7 +10,7 @@ export function getExplorerItemData(data: ExplorerItem, newThumbnails?: Record<s
 		cas_id: filePath?.cas_id || null,
 		isDir: isPath(data) && data.item.is_dir,
 		kind: (ObjectKind[objectData?.kind ?? 0] as ObjectKindKey) || null,
-		newThumb :!!newThumbnails?.[filePath?.cas_id || ''],
+		newThumb: !!newThumbnails?.[filePath?.cas_id || ''],
 		hasThumbnail: data.has_thumbnail || !!newThumbnails?.[filePath?.cas_id || ''] || false,
 		extension: filePath?.extension || null
 	};
@@ -25,8 +25,8 @@ export function getItemFilePath(data: ExplorerItem) {
 }
 
 export const SEARCH_PARAMS = z.object({
-	path: z.string().default(''),
-	limit: z.coerce.number().default(100)
+	path: z.string().optional(),
+	take: z.coerce.number().default(100)
 });
 
 export function useExplorerSearchParams() {
