+++ interface/app/$libraryId/location/$id.tsx
@@ -1,10 +1,11 @@
+import { inferQueryInput } from '@rspc/client';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { ArrowClockwise, Key, Tag } from 'phosphor-react';
 import { useEffect, useMemo } from 'react';
 import { useKey } from 'rooks';
 import { z } from 'zod';
 import {
-	ExplorerData,
+	LibraryProceduresDef,
 	useLibraryContext,
 	useLibraryMutation,
 	useRspcLibraryContext
@@ -110,33 +111,38 @@ const useToolBarOptions = () => {
 };
 
 const useItems = () => {
-	const { id: location_id } = useZodRouteParams(PARAMS);
-	const [{ path, limit }] = useExplorerSearchParams();
+	const { id: locationId } = useZodRouteParams(PARAMS);
+	const [{ path, take }] = useExplorerSearchParams();
 
 	const ctx = useRspcLibraryContext();
 	const { library } = useLibraryContext();
 
 	const explorerState = useExplorerStore();
 
+	const arg: inferQueryInput<LibraryProceduresDef, 'search.paths'> = {
+		locationId,
+		take
+	};
+
+	if (explorerState.layoutMode === 'media') arg.kind = [5, 7];
+	else arg.path = path ?? '';
+
 	const query = useInfiniteQuery({
 		queryKey: [
-			'locations.getExplorerData',
+			'search.paths',
 			{
 				library_id: library.uuid,
-				arg: {
-					location_id,
-					path: explorerState.layoutMode === 'media' ? null : path,
-					limit,
-					kind: explorerState.layoutMode === 'media' ? [5, 7] : null
-				}
+				arg
 			}
 		] as const,
-		queryFn: async ({ pageParam: cursor, queryKey }): Promise<ExplorerData> => {
-			const arg = queryKey[1];
-			(arg.arg as any).cursor = cursor;
-
-			return await ctx.client.query(['locations.getExplorerData', arg.arg]);
-		},
+		queryFn: ({ pageParam: cursor, queryKey }) =>
+			ctx.client.query([
+				'search.paths',
+				{
+					...queryKey[1].arg,
+					cursor
+				}
+			]),
 		getNextPageParam: (lastPage) => lastPage.cursor ?? undefined
 	});
