+++ interface/hooks/useZodRouteParams.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useParams } from 'react-router';
 import { z } from 'zod';
 
-export function useZodRouteParams<Z extends z.ZodType<Record<string, any>>>(schema: Z) {
+export function useZodRouteParams<Z extends z.ZodType<Record<string, any>>>(schema: Z): z.infer<Z> {
 	// eslint-disable-next-line no-restricted-syntax
 	const params = useParams();
